+++ src/main/java/com/thanh/foodOrder/dto/response/CheckOutResponseDTO.java
@@ -1,5 +1,6 @@
 package com.thanh.foodorder.dto.response;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -10,8 +11,8 @@ import lombok.Setter;
 @Setter
 public class CheckOutResponseDTO {
     private List<Long> cartDetailIds = new ArrayList<>(); // các dòng trong giỏ
-    private Double totalPrice;
-    private Double discount;
-    private Double finalPrice;
+    private BigDecimal totalPrice;
+    private BigDecimal discount;
+    private BigDecimal finalPrice;
     private Long tableId;
 }
